Don't show errors for other files by default
There's a setting to show errors for files _other_ than the files
currently open.

However, this is set to true by default as well. This results in pyright
being very confusing to developers, since it shows errors for files
_other_ than the currently open file, which doesn't really make sense.

This change makes the default what any developer would expect: when
opening a file, see errors for that file and not for other files.