Fix tsc type generation error
When building pyright-internal with `tsc --emitDeclarationOnly --declaration`, it gives an error:

```
src/analyzer/dataClasses.ts:68:16 - error TS4078: Parameter 'evaluator' of exported function has or is using private name 'TypeEvaluator'.

68     evaluator: TypeEvaluator,
                  ~~~~~~~~~~~~~
```

This is caused by a duplicate definition of `TypeEvaluator`. The definition on L119 is not exported and tsc seems to confuse it with the exported parameter on line 68, which is exported and references `TypeEvalulator` defined in typeEvaluatorTypes.ts L279.