Clarification of --version argument usage string (#5308)
* Update command-line.md

clarify that `--version` prints the version *and exits*

* Update --version usage string

Clarifies that `--version` prints the version *and exits*, this is fairly common language used by most CLI apps to indicate what the behavior of `--version` is. It may surprise new users that `--version` exits rather than appending the version with the previous language.