Added support for finding definitions of TypedDict index keys
Pulled out a helper for finding the type that the string is a key for. Copied
some logic for verifying that the StringNode is the only node within a
StringListNode from the completion logic. There appears to be some jank in the
find definition logic that doesn't verify if a StringNode is the only element
in a StringListNode. https://github.com/microsoft/pyright/issues/5655

https://github.com/microsoft/pyright/issues/5654