Fix Type Narrowing for Asymmetric Attributes After Assignment
This PR fixes the issue where an asymmetric attribute can not be narrowed after
it has been assigned in a particular scope.

Take the following for example:
```python
from typing import assert_type

class MyAttr:
   def __set__(self, instance: object, value: float) -> None:
      ...

   def __get__(self, instance: object, owner: type) -> int | None:
      ...

class MyClass:
   attr: MyAttr

MyClass.attr = 1.5

# Should be able to narrow types after initial assignment
assert MyClass.attr is not None
assert_type(MyClass.attr, int)
```

In the current version of Pyright this fails since it refuses to narrow
`MyClass.attr`. The solution to this is to save the original type of
`MyClass.attr` when an asymmetric assignment is found. This type can then be
retrieved later during flow analysis.